+++ src/domain/nina/Position.h
@@ -39,13 +39,13 @@ class Position
     /// price
     Position(const std::string& description,
              const std::string& unit,
-             float pricePerUnit,
-             float numUnits
+             double pricePerUnit,
+             double numUnits
         );
 
     /// \brief Ctor with a flat price
     Position(const std::string& description,
-             float price
+             double price
         );
 
     /// \brief Ctor only with description
@@ -58,9 +58,9 @@ class Position
   public:
     const std::string& getDescription()     const { return m_description; }
     const std::string& getUnit()            const { return m_unit; }
-    float              getPricePerUnit()    const { return m_pricePerUnit; }
-    float              getNumUnits()        const { return m_numUnits; }
-    float              getPrice()           const { return m_price; }
+    double             getPricePerUnit()    const { return m_pricePerUnit; }
+    double             getNumUnits()        const { return m_numUnits; }
+    double             getPrice()           const { return m_price; }
     bool               hasOnlyDescription() const { return m_hasOnlyDescription; }
     bool               isFreeOfCharge()     const { return m_isFreeOfCharge; }
     bool               isFlatPrice()        const { return m_isFlatPrice; }
@@ -89,9 +89,9 @@ class Position
   private:
     std::string m_description;
     std::string m_unit;
-    float       m_pricePerUnit;
-    float       m_numUnits;
-    float       m_price;
+    double      m_pricePerUnit;
+    double      m_numUnits;
+    double      m_price;
     bool        m_hasOnlyDescription;
     bool        m_isFreeOfCharge;
     bool        m_isFlatPrice;
